feat: remove state from `Command`; rename `Command` -> `Shell`
