+++ crates/z-shell/src/lib.rs
@@ -4,7 +4,6 @@ pub mod encoding;
 pub mod error;
 pub mod options;
 pub mod process;
-pub mod shell;
 
 pub use error::Error;
 type Result<T> = std::result::Result<T, Error>;
