+++ packages/desktop/src/main.rs
@@ -14,7 +14,7 @@ use tauri_plugin_shell::ShellExt;
 use tokio::{sync::mpsc, task};
 use tracing::{error, info, level_filters::LevelFilter};
 use tracing_subscriber::EnvFilter;
-use z_shell::{options::CommandOptions, shell::Shell};
+use z_shell::{options::CommandOptions, process::Shell};
 
 #[cfg(target_os = "windows")]
 use crate::common::windows::WindowExtWindows;
